requirements.txt
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
transformers==4.31.0
numpy==1.24.3
scipy==1.10.1
pyyaml==6.0
tqdm==4.65.0
openai-whisper==20231117
textgrid==1.5
cupy-cuda12x==12.0.0
h5py==3.7.0
matplotlib==3.6.2
soundfile==0.12.1
sentencepiece
python-dotenv
pyvista==0.43.1  